requirements.txt
dnspython==2.8.0
Flask==3.1.2
pymongo==4.15.1
python-dotenv==1.1.1
requests==2.32.5